# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/yardPlan": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Yard Plan"
                ],
                "summary": "Get all data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "default asc",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order By",
                        "name": "order_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.ReturnObject"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Yard Plan"
                ],
                "summary": "Create data",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "block",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "name": "container_height",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            20,
                            40
                        ],
                        "type": "integer",
                        "name": "container_size",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "row_from",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "row_to",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "slot_from",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "slot_to",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "tier_from",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "tier_to",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.ReturnObject"
                        }
                    }
                }
            }
        },
        "/yardPlan/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Yard Plan"
                ],
                "summary": "Get spesifik data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.ReturnObject"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Yard Plan"
                ],
                "summary": "Update data",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "block",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "name": "container_height",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            20,
                            40
                        ],
                        "type": "integer",
                        "name": "container_size",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "row_from",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "row_to",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "slot_from",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "slot_to",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "tier_from",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "tier_to",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.ReturnObject"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.ReturnObject": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "meta": {},
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}